Refining level up animation

+++ src/main/java/me/mrvintage/parlio/progresion/level/Skill.java
@@ -7,9 +7,9 @@ public interface Skill {
 
     void onLevelUp(ServerPlayerEntity player, int level);
 
-    int getExpReward(LevelAction action, int level);
+    double getExpReward(LevelAction action, int level);
 
-    int[] getProgression();
+    double[] getProgression();
 
     Identifier getIdentifier();
